+++ moriaengine/scripts/Basic/skills/Tracking/TrackingCharactersDialog.cs
@@ -14,6 +14,7 @@
 	Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 	Or visit http://www.gnu.org/copyleft/gpl.html
 */
+
 using System;
 using System.Collections.Generic;
 
@@ -102,7 +103,7 @@ namespace SteamEngine.CompiledScripts.Dialogs {
 				//check which character from the list is to be tracked
 				SkillSequenceArgs ssa = (SkillSequenceArgs) args[0];
 				List<AbstractCharacter> charsAround = (List<AbstractCharacter>) args[1];
-				Character charToTrack = (Character) charsAround[(int) gr.PressedButton - 10];
+				Character charToTrack = (Character) charsAround[gr.PressedButton - 10];
 
 				ssa.Target1 = charToTrack;
 				ssa.Param2 = TrackingEnums.Phase_Character_Track; //track the particular character
